Container API improvements.

+++ include/memoria/prototypes/dynvector/pages/data_page.hpp
@@ -62,7 +62,7 @@ public:
         Base::init();
     }
 
-    Short size() const {
+    Int size() const {
         return data_.size();
     }
